Use simpler test case
